Tests and node setup fixes
